eldenringapi/requirements_elden_addition.txt

# Ajout pour requirements.txt

# Elden Ring Parser dependencies
requests>=2.28.0
django-redis>=5.2.0